+++ ZXingObjCTests/oned/ZXUPCAWriterTestCase.m
@@ -24,7 +24,7 @@
   NSString *testStr = @"00010101000110110111011000100010110101111011110101010111001011101001001110110011011011001011100101000";
   ZXBitMatrix *result = [[[ZXUPCAWriter alloc] init] encode:@"485963095124"
                                                      format:kBarcodeFormatUPCA
-                                                      width:testStr.length
+                                                      width:(int)testStr.length
                                                      height:0
                                                       error:nil];
   for (int i = 0; i < testStr.length; i++) {
@@ -37,7 +37,7 @@
   NSString *testStr = @"00010100110010010011011110101000110110001010111101010100010010010001110100111001011001101101100101000";
   ZXBitMatrix *result = [[[ZXUPCAWriter alloc] init] encode:@"12345678901"
                                                      format:kBarcodeFormatUPCA
-                                                      width:testStr.length
+                                                      width:(int)testStr.length
                                                      height:0
                                                       error:nil];
   for (int i = 0; i < testStr.length; i++) {
